Added Changes for improved stability
Uproperty on all UObjects to avoid them being garbage collected early.
Rework obstacle spawning system to have the spawners check and reuse to avoid needing a reference in obsticle.
EnemyWall_BP instead of destroying itself and creating nullptr is reuse array, it instead moves so that the overlying spawner hides it and reuses it.